src/bot/telegram_bot.py
from dotenv import load_dotenv
import os
from telegram import Update
from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, CallbackContext

load_dotenv()

telegram_bot_token = os.getenv("TELEGRAM_BOT_TOKEN")
langchain_api_key = os.getenv("LANGCHAIN_API_KEY")
langchain_api_url = os.getenv("LANGCHAIN_API_URL")

# Rest of the code for the Telegram bot

class RAGQABot:
    def __init__(self, token, document_retrieval, chat_gpt, langchain_client):
        self.updater = Updater(token)
        self.document_retrieval = document_retrieval
        self.chat_gpt = chat_gpt
        self.langchain_client = langchain_client

        # Register handlers
        self.updater.dispatcher.add_handler(CommandHandler("start", self.start))
        self.updater.dispatcher.add_handler(CommandHandler("ask", self.handle_ask, pass_args=True))
        self.updater.dispatcher.add_handler(MessageHandler(Filters.text, self.handle_message))

    def start(self, update: Update, context: CallbackContext):
        welcome_message = (
            "Welcome to the RAG Q&A Bot!\n"
            "I can help you find answers to your questions based on the provided documents.\n\n"
            "To ask a question, use the /ask command followed by your question. For example:\n"
            "/ask What is the capital of France?\n\n"
            "Just send me your question, and I'll do my best to provide a helpful answer!"
        )
        update.message.reply_text(welcome_message)

    def handle_message(self, update: Update, context: CallbackContext):
        # Implement the function to handle user messages
        pass

    def handle_ask(self, update: Update, context: CallbackContext):
        if not context.args:
            update.message.reply_text("Please provide a question after the /ask command.")
            return

        question = " ".join(context.args)

        # Process the question and retrieve the answer
        # answer = ...

        update.message.reply_text(answer)

    def run(self):
        self.updater.start_polling()
        self.updater.idle()

# Instantiate the required components and create a RAGQABot instance
# document_retrieval = ...
# chat_gpt = ...
# langchain_client = ...
# bot = RAGQABot(telegram_bot_token, document_retrieval, chat_gpt, langchain_client)
# bot.run()
