src/rag_qa_system/document_retrieval.py
import faiss
import numpy as np

class DocumentRetrieval:
    def __init__(self, index_file):
        self.index = self.load_faiss_index(index_file)

    def load_faiss_index(self, index_file):
        # Implement the function to load a FAISS index
        pass

    def save_faiss_index(self, index_file):
        # Implement the function to save a FAISS index
        pass

    def add_document_vector(self, doc_id, vector):
        # Implement the function to add a document vector to the FAISS index
        pass

    def search_vectors(self, query_vector, top_k):
        # Implement the function to search the FAISS index for similar vectors
        pass
