requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
transformers
python-telegram-bot
requests
python-dotenv